Move handlers out of main.rs and start using error_chain
