+++ Cargo.toml
@@ -18,3 +18,4 @@ rand = "0.3"
 time = "0.1"
 curl = "0.4.8"
 serde_json = "1.0.4"
+error-chain = "0.11.0"
