+++ src/errors.rs
@@ -1,0 +1,1 @@
+error_chain! { }
